feat: Refactor API structure and implement new routes for community and automation management
- Introduced a new versioned API structure under `/api/v1` for better organization.
- Created routes for managing communities, including creation, retrieval, and updates.
- Implemented automation routes for creating, updating, deleting, and triggering automations.
- Added webhook handling for GitHub events, integrating Discord notifications for user contributions.
- Enhanced TypeScript schemas for community and automation data validation.
- Updated TypeScript configuration to support path mapping and type roots.
