+++ src/routes/api/v1/webhooks/routes.ts
@@ -1,0 +1,22 @@
+import { createRoute, z } from "@hono/zod-openapi";
+
+export const postGithubWebhook = createRoute({
+  method: "post",
+  path: "/github",
+  request: {
+    body: {
+      content: {
+        "application/json": {
+          schema: z.object({
+            payload: z.any(),
+          }),
+        },
+      },
+    },
+  },
+  responses: {
+    200: {
+      description: "The webhook was processed successfully",
+    },
+  },
+});
