+++ src/routes/api/v1/webhooks/schema.ts
@@ -1,0 +1,5 @@
+import { z } from "@hono/zod-openapi";
+
+export const githubWebhook = z.object({
+  payload: z.any(),
+});
